git: Create gitconfig
Configuration is pretty basic. Worth of note is that fast-forward is
disabled for merges: the expectation is that a merge commit should
always be created. If a fast-forward merge is indeed desired, it can be
achieved using the appropriate command line options.

A similar logic applies to git pull: the expectation is that it will be
fast-forwarded.