nvim: Install vim-bundler and vim-rake
The important one is vim-rake, which combines well with
vim-projectionist and also configures Vim so that 'gf' is usable on
bundler projects.

Not so sure what vim-bundler does and if it is needed at all, since
vim-rake seems to do all the work. At the same time, why not? It won't
do any harm.