Revert "fix.auto.nvim: Fix nvim_PackerSync handler"
This reverts commit ce3fdf9453c20673413855e5e97eb7d63bb76589.

The commit solved the problem it was supposed to solve, however LSP
configuration was not working properly: the on_attach local function is
not visible inside packer.nvim config block, breaking the desired
mappings.

For now, the commit is reverted in order to get work done. Before
attempting to solve this deployment issue again, a Lua LSP should be
installed for convenience and to prevent these kinds of bugs/caveats.